(MAINT) Replace arrays to make it sh-compatible
Arrays aren't supported in /bin/sh so simulate them with space-separated
strings so we can run this script on Alpine.

I removed interactive support for prompting since it was a bit more work
than refactoring the other array usages, and we don't actually need that
interactive support for the Dockerfile.